% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFuns.R
\name{chatCloud}
\alias{chatCloud}
\title{chatCloud}
\usage{
chatCloud(d, user = NULL, wordlength = 3)
}
\arguments{
\item{d}{A dataframe containing messages created from \code{\link{parseR}}}

\item{user}{The user to filter for}

\item{wordlength}{Minimum word length}
}
\description{
Plot the word frequency in users/groups chat history as a wordcloud
}
