% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiments.R
\name{chatSentiments}
\alias{chatSentiments}
\title{chatSentiments}
\usage{
chatSentiments(file_in = "data/testChat.txt", d = NA, sender = NA,
  top_sender = 5, method = "loughran")
}
\arguments{
\item{file_in}{WhatsApp chat log (if reading from file)}

\item{d}{dataframe of chat returned from \code{\link{parseR}}}

\item{sender}{The sender to filter for}

\item{top_sender}{The number of senders to show (ranked by number of words)}

\item{method}{The sentiment analysis method to use}
}
\description{
Perform a per-user sentiment analysis using
}
