% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/whatStat.R
\name{cleanTheme}
\alias{cleanTheme}
\alias{cleanTheme}
\title{cleanTheme}
\usage{
cleanTheme(base_size = 12)

cleanTheme(base_size = 12)
}
\arguments{
\item{base_size}{Set the base size for text}

\item{base_size}{Set the base size for text}
}
\description{
Clean theme for ggploting

Clean theme for ggploting
}
\keyword{clean}
