% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/whatStat.R
\name{gg_colour_hue}
\alias{gg_colour_hue}
\alias{gg_colour_hue}
\title{gg_colour_hue}
\usage{
gg_colour_hue(n)

gg_colour_hue(n)
}
\arguments{
\item{n}{Number of colours to generate}

\item{n}{Number of colours to generate}
}
\description{
Emulate the defualt ggplot colour hue

Emulate the defualt ggplot colour hue
}
\keyword{colour}
