% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/whatStat.R
\name{htmlStrip}
\alias{htmlStrip}
\alias{htmlStrip}
\title{htmlStrip}
\usage{
htmlStrip(y)

htmlStrip(y)
}
\arguments{
\item{x}{String}

\item{x}{String}
}
\description{
Strip html from string

Strip html from string
}
\keyword{html}
