% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch App in the default browser}
\usage{
launchApp()
}
\description{
Launch App in the default browser
}
\examples{
if (interactive()) {
  launchApp()
}
}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
