% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{LaunchMyApp}
\alias{LaunchMyApp}
\title{Launch App in the default browser}
\usage{
LaunchMyApp()
}
\description{
Launch App in the default browser
}
\examples{
if (interactive()) {
  LaunchMyApp()
}
}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
