% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseR.R, R/whatStat.R
\name{parseR}
\alias{parseR}
\alias{parseR}
\title{parseR}
\usage{
parseR(in_file, drop = "44", user = NA)

parseR(in_file, drop = "44", user = NA)
}
\arguments{
\item{in_file}{Whats App chat log (.txt or .zip)}

\item{user}{Select a user}

\item{in_file}{Whats App chat log (.txt or .zip)}

\item{user}{Select a user}
}
\description{
The main parsing function for extracting a dataframe from WhatsApp chat log

The main parsing function for extracting a dataframe from WhatsApp chat log
}
