% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/whatStat.R
\name{RemoveEmail}
\alias{RemoveEmail}
\alias{RemoveEmail}
\title{RemoveEmail}
\usage{
RemoveEmail(x)

RemoveEmail(x)
}
\arguments{
\item{x}{String}

\item{x}{String}
}
\description{
Strip emails addresses from string

Strip emails addresses from string
}
\keyword{emails}
