% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFuns.R
\name{senderDate}
\alias{senderDate}
\title{senderDate}
\usage{
senderDate(file_in = "data/testChat.txt", user = NA, filtYear = NA,
  d = NA)
}
\arguments{
\item{file_in}{WhatsApp chat log (if reading from file)}

\item{user}{The user to filter for}

\item{filtYear}{The year to show}

\item{d}{dataframe of chat returned from \code{\link{parseR}}}
}
\description{
Plot the date posts were sent by year
}
