% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFuns.R
\name{senderPosts}
\alias{senderPosts}
\title{senderPosts}
\usage{
senderPosts(file_in, d = NA)
}
\arguments{
\item{file_in}{WhatsApp chat log (if reading from file)}

\item{d}{dataframe of chat returned from \code{\link{parseR}}}
}
\description{
Plot the numer of posts per sender
}
