% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFuns.R
\name{senderTime}
\alias{senderTime}
\title{senderTime}
\usage{
senderTime(file_in = "data/testChat.txt", user = NA, d = NA)
}
\arguments{
\item{file_in}{WhatsApp chat log (if reading from file)}

\item{user}{The user to filter for}

\item{d}{dataframe of chat returned from \code{\link{parseR}}}
}
\description{
Plot the time posts were sent
}
