% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFuns.R
\name{wordFreq}
\alias{wordFreq}
\title{wordFreq}
\usage{
wordFreq(wordlength = 3, corpus)
}
\arguments{
\item{wordlength}{Minimum word length}

\item{corpus}{dataframe of words returned from \code{\link{makeCorpus}}}
}
\description{
Plot the word frequency in users/groups chat history
}
