% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFuns.R
\name{chatCloud}
\alias{chatCloud}
\title{chatCloud}
\usage{
chatCloud(chatdf, user = "All", wordlength = 3)
}
\arguments{
\item{user}{The user to filter for}

\item{wordlength}{Minimum word length}

\item{d}{A dataframe containing messages created from \code{\link{parseR}}}
}
\description{
Plot the word frequency in users/groups chat history as a wordcloud
}
