% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCorpus.R, R/whatStat.R
\name{makeCorpus}
\alias{makeCorpus}
\alias{makeCorpus}
\title{makeCorpus}
\usage{
makeCorpus(d)

makeCorpus(d)
}
\arguments{
\item{d}{A dataframe containing messages created from \code{\link{parseR}}}

\item{wordlength}{Minimum word length}

\item{d}{A dataframe containing messages}

\item{wordlength}{Minimum word length}
}
\description{
Make a word corpus using tm pacakge dataframe returned from \code{\link{parseR}}

Make a word corpus using tm pacakge from a datframe
}
\keyword{corpus}
